+++ tsconfig.json
@@ -34,6 +34,9 @@
       "~/data/http": [
         "src/data/http"
       ],
+      "~/infra/http": [
+        "src/infra/http"
+      ],
     },
     "allowSyntheticDefaultImports": true,
     "esModuleInterop": true,
